Deprecate old predict and refactor into packages
